+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 
  - Add ability to convert Input measurements between units ([#346](https://github.com/kizniche/mycodo/issues/346))
+ - Add unit conversions: celsius, fahrenheit, kelvin, meters, feet
  - Add ability to select whether lowering PID outputs are stored as positive or negative values
  - Add Sunrise/Sunset Conditional ([#440](https://github.com/kizniche/mycodo/issues/440))